gain/requirements.txt
pandas==1.1.4
xlrd==1.2.0
openpyxl==3.0.5
jupyterlab==2.2.9
numpy==1.19.3
tensorflow-cpu==2.3.1
tqdm==4.53.0